+++ payroll-impl/src/classification/commissioned.rs
@@ -52,7 +52,7 @@ impl PaymentClassification for CommissionedClassification {
             .sales_receipts
             .iter()
             .filter(|sr| pay_period.contains(&sr.date))
-            .fold(0 as f32, |acc, sr| {
+            .fold(0f32, |acc, sr| {
                 acc + self.calculate_pay_for_sales_receipt(sr)
             });
         debug!("commissioned_amount: {}", commissioned_amount);
